refactor: split StrategyDsl into focused files
Splits the monolithic StrategyDsl.kt and StrategyDslTest.kt into focused files based on strategy types: Time, Rollout, Filter, and List. This improves maintainability and navigation.
